toke.c: Have force_next read pl_yylval
Instead of having the caller put the token value in NEXTVAL_NEXTTOKE
(i.e., PL_nexttoke[PL_nextval]), have it put it in pl_yylval.

(pl_yylval is what is used to pass values to the parser (perly.[yc])
when yylex returns.)

PL_nexttoke[PL_nextval] won’t work any more if we make force_next do
an unshift instead of a push.  We could follow the route of the old
mad code and have all callers do start_force (effectively unshift)
before assigning to NEXTVAL_NEXTTOKE (which then points to the blank
spot at the bottom of the stack).  But it makes things simpler if the
call sites don’t have to account for this.