toke.c: Only emit POSTJOIN for ->@*
When postfix dereference syntax is interpolated in a double-quoted
string, anything ending with ->@* needs an implicit POSTJOIN token
that tells the parser to turn the preceding expression into a
join($", ...).  When I added that, I mistakenly had that token
emitted also for ->$* and ->$#*, which is unnecessary and just
gives makes more ops to execute.