regcomp.c: Make macro a lookup
The recently introduced macro isMNEMONIC_CNTRL has a look-up and several
tests in it, which occupy time and space.  Since it was only used for
debugging, that did not matter much, but future commits will use it in
more mainline code.  This commit changes it to be a single look-up,
using up one of the spare bits available for that purpose in
PL_charclass.  There are enough available bits that we aren't likely to
run out, really ever.  (We can always add a 2nd word of bits if
necessary.)