Allow for changing size of bracketed regex char class
This commit allows Perl to be compiled with a bitmap size that is larger
than 256.  This bitmap is used to directly look up whether a character
matches or not, without having to do a binary search or hash lookup.  It
might improve the performance for some installations that have a lot of
use of scripts that are above the Latin1 range.