op.c: Calculate hash for CvNAME_HEK
I assumed when I wrote that code that share_hek would calculate the
hash, like most hash functions; but this internal function assumes
the caller does it.

Hence, CVs were not sharing their heks with other types of thingies
that have heks.  A CV named foo and a GV named foo would cause two
heks of the same name to be present in the shared string table.