lib/utf8_heavy.pl: Don't return empty under miniperl
When run under miniperl before mktables has generated its tables , this
program would return empty for Unicode properties instead of an error
condition.  The problem was that when it detected this, it just
continued on to the rest of the code which assumed it had found an empty
table, instead of no table at all.  The solution is to return an e error
when one is detected.