regcomp.c: Don't prematurely skip error checking
The assertion in the comment changed by this commit was true only for
pass1 of the regex compilation; not pass2.  This makes it true in both
passes by moving it, and the code it was about past some error checking.
This error checking was executed in pass1, but not pass2.  It also
changes the warning to only be done in the second pass, part of
[perl #122671].  A future commit will fix the others