toke.c: Stop using len to indicate trailing ::
This variable stores the length of the word we are parsing.  But at
one point it starts being used as a boolean to indicate that we have a
bareword ending in two colons.  So if are looking up are bareword that
does not end in ::, we have to call strlen() to scan the string and
determine the length.