Avoid duplicate GV lookup for barewords
Since commit f74617600 (5.12), the GV lookup that this commit removes
from yylex has only been used to see whether the bareword could be a
filehandle.  The result is used by intuit_method to decide whether we
have a method call for ‘foo bar’ or ‘foo $bar’.

Doing this lookup for every bareword we encounter even when we are not
going to call intuit_method is wasteful.

The previous commit ensured that intuit_method is called only once
for each bareword, so we can put that gv lookup directly inside
intuit_method.