Fix our-sub method confusion
‘our $foo’ creates a lexical alias to a global symbol.  That lexi-
cal alias is resolved during parsing.  For instance, if you have
‘our $foo; package bar; $foo’, the last $foo is translated by the
parser into a ‘main::foo’ constant, which is then used for sym-
bol lookup.

A similar thing happens with ‘our subs’.  In
‘our sub foo; package bar; foo()’, the foo() call is first translated
into main::foo, and then there are various checks to determine how to
handle this bareword.

Sometimes it is determined to be a method call, and that’s where
things go awry.  For this name transformation should only happen when
we are going to call this sub.  If the parser concludes that it is not
actually a sub call, then the original bareword as it appeared in the
source should be used.  But that is not what was happening.  As a con-
sequence, this code compiles down to F->main::f, rather than F->f.

use experimental "lexical_subs";
our sub f;
{package F}
f F;
__END__
Undefined subroutine &main::f called at - line 4.

And that it is actually doing a method call, not just f(F) can be dem-
onstrated by the fact that extra arguments can come after F without an
intervening comma:

use experimental "lexical_subs";
our sub f { warn "@_" };
{package F}
f F "g";
__END__
F g at - line 2.

And that inheritance works:

use experimental "lexical_subs";
@ISA = "Bar";
our sub f;
undef *f;
sub Bar'f { print "bark\n" }
{package F}
f F;
__END__
bark

This commit corrects the behaviour by discarding the translated
symbol and restoring the original bareword if it turns out it is a
method name.