Change op structure for OP_METHOD and OP_METHOD_NAMED from UNOP/SVOP to METHOP. METHOP is a new structure for method ops, which holds class/method related info needed at runtime to improve perfomance of class/object method calls.
Nothing changed in functionality and/or perfomance by this commit.
It just introduces new structure which will be extended with extra
fields and used in next commits.

METHOP is binary compatible with both UNOP and SVOP, so that things
like B::Deparse and so on remain working.

Added METHOP constructors:
- newMETHOP() for method ops with dynamic method names.
  The only optype for this op is OP_METHOD.
- newMETHOP_named() for method ops with constant method names.
  Optypes for this op are: OP_METHOD_NAMED (currently) and (later)
  OP_METHOD_SUPER, OP_METHOD_REDIR, OP_METHOD_NEXT, OP_METHOD_NEXTCAN,
  OP_METHOD_MAYBENEXT

Changed perly.y, op.c and others to use METHOP instead of UNOP/SVOP