Remove GV_ADDINEVAL
This flag does absolutely nothing.  The use of 8 as a flag to pass to
gv_fetchpv was added in this commit:

commit 93233ece75d8fe98a77377e72c4d6004c5ea8691
Author: Chip Salzenberg <chip@pobox.com>
Date:   Tue Mar 3 04:39:49 1998 +0000

    [win32] merge problematic maintpatch to op.c
            #77:  "Eliminate double warnings under C<package;>"
          Files:  gv.c op.c toke.c

But nothing in gv.c made use of it at the time.

Later it was changed into a #define in commit 0f303493c6.  Subse-
quently more uses were added in 77ca0c92 and f558d5af04.

It still does nothing.  Nothing outside the perl core references it.