perldelta: Move compile-time const deref checking
This is a change in behaviour, so it should be more prominent than the
bug fixes, but it is not a backward-incompatible change (some code no
longer dies that used to), nor it is really a new feature.  Putting it
in Diagnostics is the best I can come up with.