op.c: ck_rvconst: Remove ‘double’ gv lookup
I put double in quotes because this gv lookup only happens once.

I tried commenting out the loop and leaving just the body.  All tests
passed.  Then I tried leaving the loop, but asserting that it only
iterates once.  All tests passed.

The arguments to gv_fetchsv are such that it will always add a glob
and return it, so the !gv condition always evaluates to false.  (The
glob is not added when noexpand is true, but the loop condition con-
tains !noexpand to prevent a second iteration in that case.)